create image route
- move Camera to components folder
- create submitPhoto for ocr picture
